refactor: Utilize MediaPage component in AlbumPage and ArtistPage
- Reduces code duplication by leveraging MediaPage for media details.
- Streamlines AlbumPage and ArtistPage to focus on media-specific logic.
